requirements.txt
trimesh
networkx
libigl==2.5.1
numpy
pymeshfix
gurobipy
pyvista
natsort
pyyaml
scikit-learn
robust_laplacian
potpourri3d
pymeshlab
open3d
colorcet
trame
trame-vtk
trame-vuetify
git+https://github.com/paul0noah/f2cdecimate